Un-JPEGify N14 Wood (#2227)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Redid N14 wood to not be a bunch of JPEGs. They looked atrocious.

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/user-attachments/assets/9e0cdcb2-50c6-42fd-8745-6359987ce1ff)


![image](https://github.com/user-attachments/assets/56c64c01-251e-477c-b1f8-0c3dc3192a0e)


![image](https://github.com/user-attachments/assets/d028ab32-c70b-4fed-b977-e08dcf1c2f8c)

These are poorly organized but you can tell which are new because they
don't look like rescaled jpegs

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: Zeta_Null
- tweak: Made the N14 wood types look less horrifically jpeg compressed
and like they were taken from a screenshot of a screenshot of some wood
tiles. holy fuck they were bad.